PSMNet/requirements.txt
absl-py==2.3.1
aiohappyeyeballs @ file:///home/conda/feedstock_root/build_artifacts/aiohappyeyeballs_1741775197943/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1753804962438/work
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1751626463503/work
anyio==4.12.1
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1733235340728/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1666788425425/work
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
certifi @ file:///home/task_175978575413734/conda-bld/certifi_1759786089495/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1727293517607/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
dash @ file:///home/conda/feedstock_root/build_artifacts/dash_1754906381143/work
easydict==1.13
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1739083394647/work
exceptiongroup==1.3.1
filelock==3.19.1
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1755674618058/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1755223865520/work
frozenlist @ file:///home/conda/feedstock_root/build_artifacts/frozenlist_1752167142843/work
fsspec==2025.10.0
grpcio==1.76.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1749298244729/work
hf-xet==1.2.0
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
ImageIO==2.37.2
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1748019130050/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266648/work
lazy_loader==0.4
loguru @ file:///home/conda/feedstock_root/build_artifacts/loguru_1746634721458/work
Markdown==3.9
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
msgpack @ file:///home/conda/feedstock_root/build_artifacts/msgpack-python_1749813188733/work
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1751310558090/work
munkres==1.1.4
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1755254908/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.2.1
ninja==1.13.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1732314280888/work/dist/numpy-2.0.2-cp39-cp39-linux_x86_64.whl#sha256=62d98eb3da9f13e6b227c430d01026b7427f341b3fdcb838430f2a9e520417b1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open3d-cpu @ file:///home/conda/feedstock_root/build_artifacts/open3d_1751365896967/work/build/lib/python_package/pip_package/open3d_cpu-0.19.0%2Bb53a3ad-cp39-cp39-manylinux_2_34_x86_64.whl#sha256=c05cfa016bb3418d16051818bbedfb9edf4bccafe957d7a1b6f48b43734a919b
opencv-contrib-python==4.13.0.90
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.3.3
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751482006338/work
plotly @ file:///home/conda/feedstock_root/build_artifacts/plotly_1755067229108/work
propcache @ file:///home/conda/feedstock_root/build_artifacts/propcache_1744524932661/work
protobuf==6.33.5
psutil==7.2.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1753873557/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz==2025.2
PyYAML==6.0.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
retrying @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_retrying_1754219562/work
scikit-image==0.24.0
scipy==1.13.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tifffile==2024.8.30
timm==0.4.12
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.2
transformations==2024.5.24
trimesh @ file:///home/conda/feedstock_root/build_artifacts/trimesh_1755213658857/work
triton==3.4.0
typer-slim==0.21.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.3
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692503055/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
vtk==9.4.2
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wslink @ file:///home/conda/feedstock_root/build_artifacts/wslink_1755596342224/work
xformers==0.0.32.post2
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1749554822108/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
